feat: EmailInput 수정, CommonInput 추가
