+++ src/app/auth/sign-up/page.tsx
@@ -1,4 +1,4 @@
-import AuthEmailInput from '@/components/pages/auth/AuthEmailInput';
+import SignUpForm from '@/components/pages/auth/signUp/SignUpForm';
 import AuthHeading from '@/components/ui/AuthHeading';
 import ConfirmNextButton from '@/components/ui/buttons/ConfirmNextButton.tsx';
 
@@ -9,9 +9,7 @@ export default function SignUpPage() {
         <AuthHeading>이메일과 비밀번호를</AuthHeading>
         <AuthHeading>입력해 주세요.</AuthHeading>
       </section>
-      <section className="padded space-y-6">
-        <AuthEmailInput />
-      </section>
+      <SignUpForm />
       <ConfirmNextButton text="다음" href="/" />
     </>
   );
